fix: build cli on windows
