+++ insonmnia/miner/overseer.go
@@ -10,9 +10,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/sonm-io/core/insonmnia/miner/network"
 	"github.com/sonm-io/core/insonmnia/miner/plugin"
 	"github.com/sonm-io/core/insonmnia/miner/volume"
+	"github.com/sonm-io/core/insonmnia/structs"
 	"go.uber.org/zap"
 
 	"github.com/docker/docker/api/types"
@@ -47,7 +47,7 @@ type Description struct {
 	volumes map[string]*pb.Volume
 	mounts  []volume.Mount
 
-	networks []network.Network
+	networks []structs.Network
 }
 
 func (d *Description) ID() string {
@@ -66,7 +66,7 @@ func (d *Description) GPU() bool {
 	return d.GPURequired
 }
 
-func (d *Description) Networks() []network.Network {
+func (d *Description) Networks() []structs.Network {
 	return d.networks
 }
